refactor: 이벤트와 이벤트 콜백 파일 분리, api 를 호출하는 함수를 모은 service 모듈 생성
